automatically backed up by learn

+++ index.js
@@ -26,5 +26,5 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
-  return parseInt(n)
+  return parseFloat(n)
 }
